# Refactor NatsMessage to use ByteBuffer internally.