Some groundwork for #347.