@jameshilliard 2020-10-20
Here's what I'm seeing performance wise:
Before:
```
11:25:10 AM: Executing task 'NatsAutoBench.main()'...

> Task :compileJava UP-TO-DATE
> Task :processResources NO-SOURCE
> Task :classes UP-TO-DATE

> Task :NatsAutoBench.main()
Connecting to NATS server at nats://localhost:4222
Running warmup
Current memory usage is 1019.25 mb / 1.02 gb / 16.00 gb free/total/max
Executing tests ........................................

PubOnly 0            10,000,000          5,138,146 msg/s       0.00 b/s
PubOnly 8            10,000,000          4,193,068 msg/s     31.99 mb/s
PubOnly 32           10,000,000          4,184,061 msg/s    127.69 mb/s
PubOnly 256          10,000,000          4,602,890 msg/s      1.10 gb/s
PubOnly 512          10,000,000          4,071,716 msg/s      1.94 gb/s
PubOnly 1024          1,000,000          3,113,506 msg/s      2.97 gb/s
PubOnly 4096            500,000            936,171 msg/s      3.57 gb/s
PubOnly 8192            100,000            461,962 msg/s      3.52 gb/s

PubSub 0             10,000,000          1,499,659 msg/s       0.00 b/s
PubSub 8             10,000,000          1,446,349 msg/s     11.03 mb/s
PubSub 32            10,000,000          1,330,344 msg/s     40.60 mb/s
PubSub 256           10,000,000          1,115,961 msg/s    272.45 mb/s
PubSub 512           10,000,000            981,733 msg/s    479.36 mb/s
PubSub 1024           1,000,000            839,707 msg/s    820.03 mb/s
PubSub 4096             500,000            347,289 msg/s      1.32 gb/s
PubSub 8192             100,000            144,818 msg/s      1.10 gb/s

PubDispatch 0        10,000,000          1,561,210 msg/s       0.00 b/s
PubDispatch 8        10,000,000          1,527,177 msg/s     11.65 mb/s
PubDispatch 32       10,000,000          1,414,433 msg/s     43.17 mb/s
PubDispatch 256      10,000,000          1,269,441 msg/s    309.92 mb/s
PubDispatch 512      10,000,000          1,105,003 msg/s    539.55 mb/s
PubDispatch 1024      1,000,000          1,002,108 msg/s    978.62 mb/s
PubDispatch 4096        500,000            347,199 msg/s      1.32 gb/s
PubDispatch 8192        100,000            195,271 msg/s      1.49 gb/s

ReqReply 0               20,000              9,905 msg/s       0.00 b/s
ReqReply 8               20,000              9,844 msg/s     76.91 kb/s
ReqReply 32              10,000              9,873 msg/s    308.54 kb/s
ReqReply 256             10,000              9,835 msg/s      2.40 mb/s
ReqReply 512             10,000              9,883 msg/s      4.83 mb/s
ReqReply 1024            10,000             10,172 msg/s      9.93 mb/s
ReqReply 4096            10,000              9,430 msg/s     36.84 mb/s
ReqReply 8192            10,000              9,154 msg/s     71.52 mb/s

Latency 0     5,000     39 /  52.55 / 287    +/- 0.78  (microseconds)
Latency 8     5,000     42 /  51.54 / 260    +/- 0.78  (microseconds)
Latency 32    5,000     41 /  50.87 / 167    +/- 0.76  (microseconds)
Latency 256   5,000     42 /  55.18 / 253    +/- 0.89  (microseconds)
Latency 512   5,000     43 /  59.28 / 260    +/- 0.90  (microseconds)
Latency 1024  5,000     42 /  51.17 / 187    +/- 0.75  (microseconds)
Latency 4096  5,000     42 /  52.90 / 267    +/- 0.79  (microseconds)
Latency 8192  5,000     45 /  55.60 / 859    +/- 0.85  (microseconds)

Final memory usage is 67.88 mb / 80.00 mb / 16.00 gb free/total/max

BUILD SUCCESSFUL in 2m 25s
2 actionable tasks: 1 executed, 1 up-to-date
11:27:36 AM: Task execution finished 'NatsAutoBench.main()'.

```
After:
```
11:30:12 AM: Executing task 'NatsAutoBench.main()'...

> Task :compileJava UP-TO-DATE
> Task :processResources NO-SOURCE
> Task :classes UP-TO-DATE

> Task :NatsAutoBench.main()
Connecting to NATS server at nats://localhost:4222
Running warmup
Current memory usage is 999.58 mb / 1.02 gb / 16.00 gb free/total/max
Executing tests ........................................

PubOnly 0            10,000,000          4,826,092 msg/s       0.00 b/s
PubOnly 8            10,000,000          5,019,848 msg/s     38.30 mb/s
PubOnly 32           10,000,000          4,330,835 msg/s    132.17 mb/s
PubOnly 256          10,000,000          4,942,284 msg/s      1.18 gb/s
PubOnly 512          10,000,000          4,086,727 msg/s      1.95 gb/s
PubOnly 1024          1,000,000          3,057,141 msg/s      2.92 gb/s
PubOnly 4096            500,000            916,903 msg/s      3.50 gb/s
PubOnly 8192            100,000            467,968 msg/s      3.57 gb/s

PubSub 0             10,000,000          1,533,636 msg/s       0.00 b/s
PubSub 8             10,000,000          1,433,106 msg/s     10.93 mb/s
PubSub 32            10,000,000          1,285,728 msg/s     39.24 mb/s
PubSub 256           10,000,000          1,096,267 msg/s    267.64 mb/s
PubSub 512           10,000,000            927,871 msg/s    453.06 mb/s
PubSub 1024           1,000,000            843,623 msg/s    823.85 mb/s
PubSub 4096             500,000            300,964 msg/s      1.15 gb/s
PubSub 8192             100,000            149,556 msg/s      1.14 gb/s

PubDispatch 0        10,000,000          1,303,114 msg/s       0.00 b/s
PubDispatch 8        10,000,000          1,347,450 msg/s     10.28 mb/s
PubDispatch 32       10,000,000          1,323,138 msg/s     40.38 mb/s
PubDispatch 256      10,000,000          1,157,187 msg/s    282.52 mb/s
PubDispatch 512      10,000,000          1,053,176 msg/s    514.25 mb/s
PubDispatch 1024      1,000,000            916,186 msg/s    894.71 mb/s
PubDispatch 4096        500,000            339,930 msg/s      1.30 gb/s
PubDispatch 8192        100,000            195,462 msg/s      1.49 gb/s

ReqReply 0               20,000              8,654 msg/s       0.00 b/s
ReqReply 8               20,000              8,855 msg/s     69.18 kb/s
ReqReply 32              10,000              8,619 msg/s    269.35 kb/s
ReqReply 256             10,000              8,262 msg/s      2.02 mb/s
ReqReply 512             10,000              8,927 msg/s      4.36 mb/s
ReqReply 1024            10,000              9,861 msg/s      9.63 mb/s
ReqReply 4096            10,000              9,109 msg/s     35.58 mb/s
ReqReply 8192            10,000              9,483 msg/s     74.09 mb/s

Latency 0     5,000     39 /  51.76 / 144    +/- 0.76  (microseconds)
Latency 8     5,000     39 /  50.99 / 301    +/- 0.76  (microseconds)
Latency 32    5,000     40 /  49.90 / 237    +/- 0.75  (microseconds)
Latency 256   5,000     40 /  50.53 / 255    +/- 0.78  (microseconds)
Latency 512   5,000     41 /  51.41 / 256    +/- 0.78  (microseconds)
Latency 1024  5,000     41 /  51.35 / 202    +/- 0.77  (microseconds)
Latency 4096  5,000     43 /  54.97 / 272    +/- 0.82  (microseconds)
Latency 8192  5,000     45 /  55.23 / 807    +/- 0.85  (microseconds)

Final memory usage is 67.84 mb / 80.00 mb / 16.00 gb free/total/max

BUILD SUCCESSFUL in 2m 29s
2 actionable tasks: 1 executed, 1 up-to-date
11:32:42 AM: Task execution finished 'NatsAutoBench.main()'.

```