@jameshilliard 2020-10-20
Before:
```
1:46:59 PM: Executing task 'MessageProtocolCreationBenchmark.main()'...

> Task :compileJava UP-TO-DATE
> Task :processResources NO-SOURCE
> Task :classes UP-TO-DATE
> Task :compileTestJava UP-TO-DATE
> Task :processTestResources UP-TO-DATE
> Task :testClasses UP-TO-DATE

> Task :MessageProtocolCreationBenchmark.main()
### Running benchmarks with 50,000,000 messages.

### Total time to create 50,000,000 non-utf8 messages for sending was 1,748 ms
	34.977587 ns/op
	28,589,736.279 op/sec

### Total time to create 50,000,000 utf8 messages for sending was 3,065 ms
	61.308911 ns/op
	16,310,842.564 op/sec

### Total time to create 50,000,000 a protocol message was 913 ms
	18.262107 ns/op
	54,758,195.611 op/sec

BUILD SUCCESSFUL in 6s
4 actionable tasks: 1 executed, 3 up-to-date
1:47:05 PM: Task execution finished 'MessageProtocolCreationBenchmark.main()'.
```

Benchmarks after further optimization of `fastUtf8Length`:
```
1:47:11 PM: Executing task 'MessageProtocolCreationBenchmark.main()'...

> Task :compileJava UP-TO-DATE
> Task :processResources NO-SOURCE
> Task :classes UP-TO-DATE
> Task :compileTestJava UP-TO-DATE
> Task :processTestResources UP-TO-DATE
> Task :testClasses UP-TO-DATE

> Task :MessageProtocolCreationBenchmark.main()
### Running benchmarks with 50,000,000 messages.

### Total time to create 50,000,000 non-utf8 messages for sending was 1,870 ms
	37.406164 ns/op
	26,733,561.637 op/sec

### Total time to create 50,000,000 utf8 messages for sending was 3,242 ms
	64.849350 ns/op
	15,420,354.966 op/sec

### Total time to create 50,000,000 a protocol message was 393 ms
	7.870244 ns/op
	127,060,856.575 op/sec

BUILD SUCCESSFUL in 5s
4 actionable tasks: 1 executed, 3 up-to-date
1:47:17 PM: Task execution finished 'MessageProtocolCreationBenchmark.main()'.
```